Feature/qemu (#290)
* Add toolchain file for Cortex M3

* Add CMAKE_CROSS_COMPILING_EMULATOR

* Do not use BSD specific random functions.

* Add toolchain file for arm linux.

* Correct unit tests for endianes checks.

* Correct UTF-8 checker for big endian machines.

* Cast timeout values correctly for 32 bit machines.

* Let the http parser know EOF.

* Make content-length a size_t.

* Make websocket frame-length a size_t.

* Add toolchain file for powerpc.

* Use bsd random function if applicable.

* Remove redundant cast.

* Set close_on_exit flag.

* Provide correct casts.

+++ src/cio_http_client.h
@@ -71,7 +71,7 @@ struct cio_http_client_private {
 	struct cio_timer request_timer;
 	struct cio_timer response_timer;
 
-	uint64_t remaining_content_length;
+	size_t remaining_content_length;
 	bool should_keepalive;
 	bool close_immediately;
 	bool headers_complete;
@@ -186,7 +186,7 @@ struct cio_http_client {
 	 * @ref cio_http_location_handler_on_headers_complete "on_headers_complete callback" you can rely
 	 * on the field being set.
 	 */
-	uint64_t content_length;
+	size_t content_length;
 
 	/**
 	 * @brief The HTTP method (i.e. GET, POST, PUT, ...) of the HTTP client request.
